central/requirements.txt
django==1.10.4
psycopg2
djangorestframework==3.4.7
django-rest-swagger==0.3.8
drf-extensions==0.2.8
PyJWT==1.4.2
pytz
python-dateutil==2.2
requests
django-cors-middleware==1.3.1
boto3==1.4.0
Pillow
bitarray
pdfrw==0.3
reportlab==2.7